Improve coprocess tests, randomize grpc port with :0 to prevent collisions
